+++ Part One/2_Result/SelfDrivingCar.World/Road.cs
@@ -3,5 +3,5 @@ namespace SelfDrivingCar.World;
 public class Road {
   public double Distance { get; set; }
   public double Bearing { get; set; }
-  public int SpeedLimit { get; set; }
+  public double SpeedLimit { get; set; }
 }
