+++ Part Two/1_NavigationFailure/SelfDrivingCar.Car/InternalTools/SignReader.cs
@@ -4,7 +4,7 @@ namespace SelfDrivingCar.Car.InternalTools;
 
 public class SignReader
 {
-    public int GetSpeedForCurrentRoad(Road road)
+    public double GetSpeedForCurrentRoad(Road road)
     {
         return road.SpeedLimit;
     }
